CGP-56  #SB  Added a bug fix for when trying to update a Track and it sets the duration back to 0. Changed the duration variable from int to big Long using Gia and Cameron's idea, changed the data type also in the Track Service test

+++ src/main/java/com/qa/choonz/persistence/domain/Track.java
@@ -34,14 +34,14 @@ public class Track {
     private Playlist playlist;
 
     // in seconds
-    private int duration;
+    private Long duration;
     
     @Size(max = 5000)
     private String lyrics;
 
  
 
-	public Track(@NotNull @Size(max = 100) String name, int duration, @Size(max = 5000) String lyrics) {
+	public Track(@NotNull @Size(max = 100) String name, Long duration, @Size(max = 5000) String lyrics) {
 		super();
 		this.name = name;
 		this.duration = duration;
@@ -50,7 +50,7 @@ public class Track {
 
 
 
-	public Track(Long id, @NotNull @Size(max = 100) String name, int duration, @Size(max = 5000) String lyrics) {
+	public Track(Long id, @NotNull @Size(max = 100) String name, Long duration, @Size(max = 5000) String lyrics) {
 		super();
 		this.id = id;
 		this.name = name;
